backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-hcl2==4.3.2
pydantic==2.7.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
redis==5.0.4
anthropic==0.28.0
python-multipart==0.0.9
httpx==0.27.0
boto3==1.34.0
celery==5.4.0
alembic==1.13.1
